[refactor] 바의 길이가 constant가 아닌 비율대로 나오도록 기능 수정함 	1. 최대 오염 값을 200으로 두고 200보다 크거나 같으면 비율을 1로, 200보다 작으면 dustValue/200 로 비율을 둠 	2. 이를 위해 이전의 constraint를 지우기 위해 removeConstraint를 사용함 	3. 값이 유연하게 적용되독 dustValue의 타입을 Double로 바꿈

+++ iOS/DustNoticeApp/DustNoticeApp/Models/DustInfo.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct DustInfo: Codable {
-    let dustValue: UInt
+    let dustValue: Double
     let dustGrade: UInt
     let measureDate: Date
